Remove xfel from optional_modules in libtbx_config
If dials is configured before cctbx_project/xfel, then xfel is inserted in
front of dials in the SCons configuration order, because libtbx.env_config
doesn't topologically sort, but instead only evaluates dependency order for
packages that are not already in the module list.

This means that passing the order "dials [xfel]" to libtbx.configure puts xfel
before dials in the module order/SConstruct, and the order "xfel [dials]" puts
dials before xfel in the module/SConstruct order. Both of these modules
currently have a circular dependency relationship.

xfel before dials currently works, because xfel assumes that it is first and
overwrites the env_etc utility variables, but this has the effect that it is
impossible for dials to change the way any of this is resolved.

This change removes the "optional" dependency, meaning that if dials is
configured, libtbx doesn't automatically attempt to configure the xfel module
first.

While here, tidy up the old python_required and conda_required entries, as we
now manage dependencies through separate package lists in .conda-envs/